Refactor theme into separate files
move box shadows

Fix theme
