+++ src/theme/focus.ts
@@ -1,0 +1,25 @@
+import { CSSObject } from 'styled-components'
+
+import { asElementTypes } from '../utils/asElementTypes'
+
+import { boxShadows } from './boxShadows'
+
+import { borderWidths } from './borders'
+
+export const focusPartials = asElementTypes<CSSObject>()({
+  default: {
+    outline: 'none',
+    boxShadow: boxShadows.focused,
+  },
+  insetAbsolute: {
+    outline: 'none',
+    position: 'absolute',
+    content: "''",
+    pointerEvents: 'none',
+    top: `${borderWidths.focus}px`,
+    right: `${borderWidths.focus}px`,
+    left: `${borderWidths.focus}px`,
+    bottom: `${borderWidths.focus}px`,
+    boxShadow: boxShadows.focused,
+  },
+})
