+++ src/theme/fonts.ts
@@ -1,0 +1,6 @@
+export const fontFamilies = {
+  mono: '"Monument Mono", monospace',
+  semi: '"Monument Semi-Mono", "Monument", "Inter", "Helvetica", "Arial", "sans-serif"',
+  sans: '"Inter", "Helvetica", "Arial", "sans-serif"',
+  sansHero: '"Monument", "Inter", "Helvetica", "Arial", "sans-serif"',
+}
