+++ src/theme/zIndexes.ts
@@ -1,0 +1,6 @@
+export const zIndexes = {
+  base: 0,
+  selectPopover: 500,
+  modal: 1000,
+  tooltip: 2000,
+}
