+++ src/utils/asElementTypes.tsx
@@ -1,0 +1,7 @@
+export function asElementTypes<T>() {
+  return function ret<Obj>(obj: {
+      [K in keyof Obj]: T
+    }) {
+    return obj
+  }
+}
